requirements.txt
requests
beautifulsoup4
psycopg2-binary
python-telegram-bot[all]
python-dotenv
schedule
pandas
pathlib